Use helpers on shared methods in rspec test suite
Little refactoring to externalise the logic of `post_with_params` in a shared helper
